Remove unused imports and move print to GeneralUtilities
